+++ tests/test_pipelines.py
@@ -1500,6 +1500,57 @@ class TestPostgreSQLPipeline:
         self.pipeline.db_cursor.execute("select * from race_payoff")
         assert len(self.pipeline.db_cursor.fetchall()) == 0
 
+    def test_process_race_payoff_item_11(self):
+        # Setup
+        item = RacePayoffItem()
+        item['favorite'] = ['-']
+        item['horse_number'] = ['4']
+        item['odds'] = ['150円']
+        item['payoff_type'] = ['単勝']
+        item['race_id'] = ['sponsorCd=29&raceDy=20171023&opTrackCd=55&raceNb=10']
+
+        # Before check
+        self.pipeline.db_cursor.execute("select * from race_payoff")
+        assert len(self.pipeline.db_cursor.fetchall()) == 0
+
+        # Execute
+        new_item = self.pipeline.process_item(item, None)
+
+        # Check return
+        assert new_item['race_payoff_id'] == 'sponsorCd=29&raceDy=20171023&opTrackCd=55&raceNb=10_win_4'
+        assert new_item['race_id'] == 'sponsorCd=29&raceDy=20171023&opTrackCd=55&raceNb=10'
+        assert new_item['payoff_type'] == 'win'
+        assert new_item['horse_number_1'] == 4
+        assert new_item['horse_number_2'] is None
+        assert new_item['horse_number_3'] is None
+        assert new_item['odds'] == 1.5
+        assert new_item['favorite'] is None
+
+        # Check db
+        self.pipeline.db_cursor.execute("select * from race_payoff")
+
+        race_payoffs = self.pipeline.db_cursor.fetchall()
+        assert len(race_payoffs) == 1
+
+        race_payoff = race_payoffs[0]
+        assert race_payoff['race_payoff_id'] == 'sponsorCd=29&raceDy=20171023&opTrackCd=55&raceNb=10_win_4'
+        assert race_payoff['race_id'] == 'sponsorCd=29&raceDy=20171023&opTrackCd=55&raceNb=10'
+        assert race_payoff['payoff_type'] == 'win'
+        assert race_payoff['horse_number_1'] == 4
+        assert race_payoff['horse_number_2'] is None
+        assert race_payoff['horse_number_3'] is None
+        assert race_payoff['odds'] == 1.5
+        assert race_payoff['favorite'] is None
+
+        # Execute (2)
+        self.pipeline.process_item(item, None)
+
+        # Check db (2)
+        self.pipeline.db_cursor.execute("select * from race_payoff")
+
+        race_payoffs = self.pipeline.db_cursor.fetchall()
+        assert len(race_payoffs) == 1
+
     def test_process_horse_item_1(self):
         # Setup
         item = HorseItem()