+++ investment_local_horse_racing_crawler/scrapy/contracts.py
@@ -1,12 +1,12 @@
-import logging
-
 from scrapy.contracts import Contract
 from scrapy.exceptions import ContractFail
 from scrapy.http import Request
 
-from investment_local_horse_racing_crawler.items import OddsWinPlaceItem, RaceResultItem, RacePayoffItem, HorseItem, JockeyItem, TrainerItem
+from investment_local_horse_racing_crawler.scrapy.items import OddsWinPlaceItem, RaceResultItem, RacePayoffItem, HorseItem, JockeyItem, TrainerItem
+from investment_local_horse_racing_crawler import app_logging
 
-logger = logging.getLogger(__name__)
+
+logger = app_logging.get_logger()
 
 
 class ScheduleListContract(Contract):
@@ -35,11 +35,31 @@ class ScheduleListContract(Contract):
             if r.url.startswith("https://www.oddspark.com/keiba/RaceRefund.do?"):
                 continue
 
+            if r.url.startswith("https://www.oddspark.com/keiba/OneDayRaceList.do?"):
+                continue
+
             raise ContractFail(f"Unknown url: {r.url}")
 
 
 class RaceRefundListContract(Contract):
     name = "race_refund_list"
+
+    def post_process(self, output):
+        # Check requests
+        requests = [o for o in output if isinstance(o, Request)]
+        if len(requests) == 0:
+            raise ContractFail("Empty requests")
+
+        # Check unknown url
+        for r in requests:
+            if r.url.startswith("https://www.oddspark.com/keiba/RaceList.do?"):
+                continue
+
+            raise ContractFail(f"Unknown url: {r.url}")
+
+
+class OneDayRaceListContract(Contract):
+    name = "one_day_race_list"
 
     def post_process(self, output):
         # Check requests