+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unrelease]
+### Added
+- [#8166: キャッシュ有効期限をページによって変える](https://redmine.u6k.me/issues/8166)
+
 ## [1.0.3] - 2020-04-09
 ### Fixed
 - [#8148: 出馬表のパースに失敗した…馬体重などが空](https://redmine.u6k.me/issues/8148)